COOL: Crossbreed C++ coroutines and our tasks
We were reluctant to introduce the coroutine-based approach
as it used to look less efficient than scheduling a (mostly) static
graph of tasks. It turns out that dynamic subflow/subchain construction
is required even for some relatively simple problems, and the
coroutine-based approach is the cleanest solution in these cases.